refactor: initial removal of discovery code from net move discovery to its own folder and files
