+++ crates/networking/p2p/discv4/helpers.rs
@@ -1,0 +1,31 @@
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+//TODO remove unwraps
+pub fn get_expiration(seconds: u64) -> u64 {
+    (SystemTime::now() + Duration::from_secs(seconds))
+        .duration_since(UNIX_EPOCH)
+        .unwrap()
+        .as_secs()
+}
+
+pub fn is_expired(expiration: u64) -> bool {
+    // this cast to a signed integer is needed as the rlp decoder doesn't take into account the sign
+    // otherwise a potential negative expiration would pass since it would take 2^64.
+    (expiration as i64)
+        < SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_secs() as i64
+}
+
+pub fn time_since_in_hs(time: u64) -> u64 {
+    let time = SystemTime::UNIX_EPOCH + std::time::Duration::from_secs(time);
+    SystemTime::now().duration_since(time).unwrap().as_secs() / 3600
+}
+
+pub fn time_now_unix() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap()
+        .as_secs()
+}
