+++ crates/networking/p2p/discv4/messages.rs
@@ -1,5 +1,3 @@
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
 use crate::types::{Endpoint, Node, NodeRecord};
 use bytes::BufMut;
 use ethrex_core::{H256, H512, H520};
@@ -12,35 +10,7 @@ use ethrex_rlp::{
 use k256::ecdsa::{RecoveryId, Signature, SigningKey, VerifyingKey};
 use sha3::{Digest, Keccak256};
 
-//todo add tests
-pub fn get_expiration(seconds: u64) -> u64 {
-    (SystemTime::now() + Duration::from_secs(seconds))
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_secs()
-}
-
-pub fn is_expired(expiration: u64) -> bool {
-    // this cast to a signed integer is needed as the rlp decoder doesn't take into account the sign
-    // otherwise a potential negative expiration would pass since it would take 2^64.
-    (expiration as i64)
-        < SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_secs() as i64
-}
-
-pub fn time_since_in_hs(time: u64) -> u64 {
-    let time = SystemTime::UNIX_EPOCH + std::time::Duration::from_secs(time);
-    SystemTime::now().duration_since(time).unwrap().as_secs() / 3600
-}
-
-pub fn time_now_unix() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_secs()
-}
+use super::helpers::time_now_unix;
 
 #[derive(Debug, PartialEq)]
 pub enum PacketDecodeErr {
